MagicCombo: Magic combination handler
Whenever a special combination is held down at the same time, run a
custom callback function. This makes it possible to run any action in
response to a magic combination.

Takes a dictionary as its only argument, which describes (in a bit
field) which keys held will result in what action.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>