SpaceCadet: Major simplification
There is only one mode now: holding. If either shift is pressed, they
register immediately. If they are released before a timeout, without any
interrupting key, then parens are registered too. If they are released
after the timeout, without interruption, the parens are not registered.
If they are released at any point, but after interruption, parens are
not registered, either.

This makes the code much simpler.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>