Unicode: Helpers for Unicode input
These are low-level things one can use in macros and similar, to input
Unicode symbols. Since the way this can be done varies from OS to OS,
the plugin hides most of that, and we only have to select the OS.

The selected OS gets saved to EEPROM, unless it is set to "auto", in
which case we'll try to guess. Users are encouraged to manually select
an input method.

Fixes #55.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>