Try to avoid implying this is Model 01 specific
Most of the stuff here would work with any board that KeyboardioFirmware
supports. Even though that's only the Model 01 now, that may not remain
the case. For this reason, remove most things that would imply that the
library is specific to the Model 01, and rather, say it's for any
Keyboardio board.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>