Move all configuration to a configure method
All of the plugins now create a static object by default, and all
configuration has been moved to `configure` methods. This makes it
easier to use the plugins when no additional configuration is needed,
and also frees the user from the burden of having to find a variable
name.

That, in turn, allows some of the thins to be simplified, like the
`tapDanceActionKeys` macro, that no longer requires an object argument:
it will use `TapDance`.

Fixes #81.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>