SpaceCadet: Make it work
First, initialize the `leftParen`/`rightParen` variables, so we have a
reasonable default. Then, when sending the key, make sure that the appropriate
`Shift` is pressed too, by injecting it.

And since we need to press keys, and then release them too, we need to trigger a
`Keyboard.sendReport` as well.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>