OneShot: Cancel the one-shot when the interrupter is released
If we cancel when the interrupter is pressed, there's the problem of held keys:
a single tap usually produces at least one hold event too. We do not want the
hold to fire any other actions, so we cancel the one-shot effect on keyup.

Not the best solution, but the simplest working one so far.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>