OneShot: Correct the timeout handling
Fixed the `hasTimeOut` macro, so it has a chance to fire. Also made it so that
the timer is cleared only when we did cancel any one-shots, not always when we
should have.

This in turn fixes the case of a long-pressed one-shot key not timing out
properly when released.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>